Revert back to React frontend from JavaFX
