--- music-client/build.gradle
@@ -1,29 +1,0 @@
-plugins {
-  id 'application'
-  id 'org.openjfx.javafxplugin' version '0.0.9'
-    id 'org.jetbrains.kotlin.jvm' version '1.4.31'
-}
-
-repositories {
-    mavenCentral()
-}
-
-javafx {
-    version = "15.0.1"
-    modules = [ 'javafx.controls' ]
-}
-
-mainClassName = 'com.benbasten.groupmusicserver.GroupMusicClientApplication'
-dependencies {
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8"
-}
-compileKotlin {
-    kotlinOptions {
-        jvmTarget = "1.8"
-    }
-}
-compileTestKotlin {
-    kotlinOptions {
-        jvmTarget = "1.8"
-    }
-}
