--- music-server/src/main/kotlin/com/benbasten/groupmusicserver/controller/MusicController.kt
@@ -1,20 +1,0 @@
-package com.benbasten.groupmusicserver.controller
-
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RestController
-
-@RestController
-@RequestMapping("/api/gms")
-class MusicController {
-
-    @GetMapping("/hello")
-    fun hello(): String {
-        return "Hello World!"
-    }
-
-    @GetMapping("/music")
-    fun getMusicListings(): List<String> {
-        return emptyList()
-    }
-}
