+++ settings.gradle.kts
@@ -1,2 +1,1 @@
 rootProject.name = "group-music-server"
-include("music-server", "music-client")
